Port HtmlBinding to Binding API #35
