+++ main.js
@@ -11,7 +11,7 @@ import ViewPort from './lib/ViewPort.js';
 import * as Config from './lib/Config';
 import './lib/bindings/IfBinding.js';
 import './lib/ElementToScopeBinding.js';
-import './lib/HtmlBinding.js';
+import './lib/bindings/HtmlBinding.js';
 import './lib/bindings/CloakBinding.js';
 import './lib/bindings/AttributeBinding';
 import './lib/bindings/AnimationBinding';
